web-animations/animation-model/keyframe-effects/opacity-clamping.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>opacity</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>

<div id="log"></div>

<script>

test(t => {
  const element = createDiv(t);
  const frames = [{ opacity:-1},{opacity:-1}];
  const timing = {
    duration: Infinity,
    easing: "linear",
    fill: "backwards",
    composite: "add"
  };
  const effect = new KeyframeEffect(element, frames, timing);
  const animation = new Animation(effect, document.timeline);
  animation.play();
  const actual = Number(getComputedStyle(element).getPropertyValue("opacity"));
  const expected = 0;
  animation.cancel();
  assert_equals(actual, expected);
}, "opacity should allow negative number values in keyframes");

test(t => {
  const element = createDiv(t);
  const frames = [{ opacity:-2},{opacity:-2}];
  const timing = {
    duration: Infinity,
    easing: "linear",
    fill: "backwards",
    composite: "add"
  };
  const effect = new KeyframeEffect(element, frames, timing);
  const animation = new Animation(effect, document.timeline);
  animation.play();
  const actual = Number(getComputedStyle(element).getPropertyValue("opacity"));
  const expected = 0;
  animation.cancel();
  assert_equals(actual, expected);
}, "opacity should allow negative number values in keyframes and clamp after additive compositing");

test(t => {
  const element = createDiv(t);
  const frames = [{ opacity:"-100%"},{opacity:"-100%"}];
  const timing = {
    duration: Infinity,
    easing: "linear",
    fill: "backwards",
    composite: "add"
  };
  const effect = new KeyframeEffect(element, frames, timing);
  const animation = new Animation(effect, document.timeline);
  animation.play();
  const actual = Number(getComputedStyle(element).getPropertyValue("opacity"));
  const expected = 0;
  animation.cancel();
  assert_equals(actual, expected);
}, "opacity should allow negative percentage values in keyframes");

test(t => {
  const element = createDiv(t);
  const frames = [{ opacity:"-200%"},{opacity:"-200%"}];
  const timing = {
    duration: Infinity,
    easing: "linear",
    fill: "backwards",
    composite: "add"
  };
  const effect = new KeyframeEffect(element, frames, timing);
  const animation = new Animation(effect, document.timeline);
  animation.play();
  const actual = Number(getComputedStyle(element).getPropertyValue("opacity"));
  const expected = 0;
  animation.cancel();
  assert_equals(actual, expected);
}, "opacity should allow negative percentage values in keyframes and clamp after additive compositing");


test(t => {
  const element = createDiv(t);
  element.style.opacity = -1; // this is a valid <number>
  const frames = [{ opacity:2},{opacity:2}];
  const timing = {
    duration: Infinity,
    easing: "linear",
    fill: "backwards",
    composite: "add"
  };
  const effect = new KeyframeEffect(element, frames, timing);
  const animation = new Animation(effect, document.timeline);
  animation.play();
  const actual = Number(getComputedStyle(element).getPropertyValue("opacity"));
  const expected = 1;
  animation.cancel();
  assert_equals(actual, expected);
}, "opacity should allow number values greater than one in keyframes");

test(t => {
  const element = createDiv(t);
  element.style.opacity = 0;
  const frames = [{ opacity:2},{opacity:2}];
  const timing = {
    duration: Infinity,
    easing: "linear",
    fill: "backwards",
    composite: "add"
  };
  const effect = new KeyframeEffect(element, frames, timing);
  const animation = new Animation(effect, document.timeline);
  animation.play();
  const actual = Number(getComputedStyle(element).getPropertyValue("opacity"));
  const expected = 1;
  animation.cancel();
  assert_equals(actual, expected);
}, "opacity should allow number values greater than one in keyframes and clamp after additive compositing");

test(t => {
  const element = createDiv(t);
  element.style.opacity = -1; // this is a valid <number>
  const frames = [{ opacity:"200%"},{opacity:"200%"}];
  const timing = {
    duration: Infinity,
    easing: "linear",
    fill: "backwards",
    composite: "add"
  };
  const effect = new KeyframeEffect(element, frames, timing);
  const animation = new Animation(effect, document.timeline);
  animation.play();
  const actual = Number(getComputedStyle(element).getPropertyValue("opacity"));
  const expected = 1;
  animation.cancel();
  assert_equals(actual, expected);
}, "opacity should allow percentage values greater than 100 in keyframes");

test(t => {
  const element = createDiv(t);
  element.style.opacity = 0;
  const frames = [{ opacity:"200%"},{opacity:"200%"}];
  const timing = {
    duration: Infinity,
    easing: "linear",
    fill: "backwards",
    composite: "add"
  };
  const effect = new KeyframeEffect(element, frames, timing);
  const animation = new Animation(effect, document.timeline);
  animation.play();
  const actual = Number(getComputedStyle(element).getPropertyValue("opacity"));
  const expected = 1;
  animation.cancel();
  assert_equals(actual, expected);
}, "opacity should allow percentage values greater than 100 in keyframes and clamp after additive compositing");

</script>
